// client/src/interfaces/Figure.ts
export interface Figure {
    id: string
    method: string
    figure: {
        color: string
        h: number;
        lineWidth: number;
        strokeColor: string;
        strokeStyle: string
        type: string;
        w: number;
        x: number;
        y: number;
        r: number
    }
}